push

+++ src/main/java/org/ParkingLotWithDesign/Strategy/CreditCardPaymentStrategy.java
@@ -3,7 +3,7 @@ package org.ParkingLotWithDesign.Strategy;
 public class CreditCardPaymentStrategy implements PaymentStrategy {
 
     @Override
-    public void pay(long amount) {
+    public void pay(double amount) {
         System.out.println("Credit card payment is done ");
     }
 }
