+++ src/main/java/org/ParkingLotWithDesign/Strategy/DebitCardPaymentStrategy.java
@@ -2,7 +2,7 @@ package org.ParkingLotWithDesign.Strategy;
 
 public class DebitCardPaymentStrategy implements PaymentStrategy{
     @Override
-    public void pay(long amount) {
+    public void pay(double amount) {
         System.out.println("Debit card payment is done ");
     }
 }
