+++ src/main/java/org/ParkingLotWithDesign/Strategy/GooglePaymentStrategy.java
@@ -2,7 +2,7 @@ package org.ParkingLotWithDesign.Strategy;
 
 public class GooglePaymentStrategy implements PaymentStrategy {
     @Override
-    public void pay(long amount) {
+    public void pay(double amount) {
         System.out.println("Payment using google pay");
     }
 }
